isalpha, strlcat, strnstr, strrchr, strchr all finished

+++ ft_memchr.c
@@ -6,7 +6,7 @@
 /*   By: hde-vos <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/05/23 17:03:25 by hde-vos           #+#    #+#             */
-/*   Updated: 2019/05/24 13:12:28 by hde-vos          ###   ########.fr       */
+/*   Updated: 2019/05/24 14:58:05 by hde-vos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	*ft_memchr(const void *s, int c, size_t n)
 	str = s;
 	i = 0;
 	x = c;
-	while (str[i] != '\0' && i < n)
+	while (str[i] != '\0' && i < (int)n)
 	{
 		if (str[i] == x)
 			return ((void *)s + i);
